# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Customer Controller
  description: ""
  termsOfService: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Joel Silva
    email: joeltadeu@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
  - url: http://localhost:9081
    description: Generated server url
...
paths:
...
  /v1/customers/{customerId}/emails/{id}:
    get:
      tags:
        - Email Controller
      summary: Retrieve a customer email by id
      description: This endpoint is responsible to retrieve the customer email data
        by id
      operationId: findEmailById
      parameters:
        - name: customerId
          in: path
          description: Id of the customer
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: id
          in: path
          description: Id of the email to be searched
          required: true
          schema:
            type: integer
            format: int64
          example: 15
      responses:
        "200":
          description: Return email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailResponse'
        "500":
          description: An unexpected error occurred during retrieve the customer email
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to retrieve a customer email
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
        "404":
          description: Email not found
          content:
            application/json:
              examples:
                Not found example:
                  description: A not found response example when trying to retrieve
                    a customer email does not exist
                  value:
                    code: 404
                    status: Not Found
                    description: Email with id 4 was not found
                    date: 2022-10-28T21:39:54.603263862
    put:
      tags:
        - Email Controller
      summary: Update a customer email
      description: This endpoint is responsible to update a customer email data by
        id
      operationId: updateEmail
      parameters:
        - name: customerId
          in: path
          description: Id of the customer
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: id
          in: path
          description: Id of the email to be updated
          required: true
          schema:
            type: integer
            format: int64
          example: 15
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRequest'
        required: true
      responses:
        "200":
          description: Email updated
          content:
            application/json: {}
        "500":
          description: An unexpected error occurred during update the customer email
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to update a customer email
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
        "404":
          description: Email not found
          content:
            application/json:
              examples:
                Not found example:
                  description: A not found response example when trying to update
                    a customer email does not exist
                  value:
                    code: 404
                    status: Not Found
                    description: Email with id 4 was not found
                    date: 2022-10-28T21:39:54.603263862
        "400":
          description: Email request is invalid
          content:
            application/json:
              examples:
                Bad request example:
                  description: A bad request response example when trying to update
                    a customer email
                  value: |
                    {
                         "code": 400,
                         "status": "Bad Request",
                         "description": "Validation Exception",
                         "date": "2022-11-25T20:14:04.078313400",
                         "attributes": [
                             {
                                 "attribute": "type",
                                 "errors": [
                                     "Email Type cannot be null"
                                 ]
                             },
                             {
                                 "attribute": "email",
                                 "errors": [
                                     Email cannot be null"
                                 ]
                             }
                         ]
                    }
    delete:
      tags:
        - Email Controller
      summary: Delete the customer email by id
      description: This endpoint is responsible to delete the customer email by id
      operationId: deleteEmail
      parameters:
        - name: customerId
          in: path
          description: Id of the customer
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: id
          in: path
          description: Id of the email to be deleted
          required: true
          schema:
            type: integer
            format: int64
          example: 15
      responses:
        "204":
          description: Email deleted
          content:
            application/json: {}
        "500":
          description: An unexpected error occurred during delete the customer email
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to delete customer email
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
        "404":
          description: Email not found
          content:
            application/json:
              examples:
                Not found example:
                  description: A not found response example when trying to delete
                    a customer email does not exist
                  value:
                    code: 404
                    status: Not Found
                    description: Email with id 4 was not found
                    date: 2022-10-28T21:39:54.603263862
...
  /v1/customers/{customerId}/emails:
    get:
      tags:
        - Email Controller
      summary: Retrieve a list of emails by customer id
      description: This endpoint is responsible to retrieve a list of customer emails
        based on the applied filters
      operationId: listAllEmails
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: filter
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/EmailFilter'
      responses:
        "500":
          description: An unexpected error occurred during retrieve the customer emails
            list
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to retrieve a customer emails list
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
        "200":
          description: Return customer emails list
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageEmailResponse'
    post:
      tags:
        - Email Controller
      summary: Create a email
      description: This endpoint is responsible to create a new email
      operationId: createEmail
      parameters:
        - name: customerId
          in: path
          description: Id of the customer
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRequest'
        required: true
      responses:
        "201":
          description: Email created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailResponse'
        "400":
          description: Email request is invalid
          content:
            application/json:
              examples:
                Bad request example:
                  description: A bad request response example when trying to create
                    an email
                  value: |
                    {
                         "code": 400,
                         "status": "Bad Request",
                         "description": "Validation Exception",
                         "date": "2022-11-25T20:14:04.078313400",
                         "attributes": [
                             {
                                 "attribute": "type",
                                 "errors": [
                                     "Email Type cannot be null"
                                 ]
                             },
                             {
                                 "attribute": "email",
                                 "errors": [
                                     Email cannot be null"
                                 ]
                             }
                         ]
                    }
        "500":
          description: An unexpected error occurred during creating the email
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to create an email
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
...
components:
  schemas:
...
    EmailRequest:
      required:
        - email
        - type
      type: object
      properties:
        type:
          type: string
          description: Email type
          example: PERSONAL
          enum:
            - WORK
            - PERSONAL
        email:
          type: string
          description: Email
          example: damian.obrien@gmail.com
...
    EmailResponse:
      type: object
      properties:
        id:
          type: integer
          description: Email Id
          format: int64
          example: 13
        type:
          type: string
          description: Email type
          example: PERSONAL
          enum:
            - WORK
            - PERSONAL
        email:
          type: string
          description: Email
          example: damian.obrien@gmail.com
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        paged:
          type: boolean
        unpaged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        unsorted:
          type: boolean
        sorted:
          type: boolean
    EmailFilter:
      required:
        - pageNumber
        - pageSize
      type: object
      properties:
        pageNumber:
          type: integer
          description: Zero-base page index (0..N)
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: Number of records per page
          format: int32
          example: 10
          default: 10
        sortDirection:
          type: string
          description: Sort direction
          example: ASC|DESC
          enum:
            - ASC
            - DESC
        sortBy:
          type: string
          description: Sort column name
          example: firstName
        type:
          type: string
          description: Email type to be searched
          example: PERSONAL|WORK
          enum:
            - WORK
            - PERSONAL
        email:
          type: string
          description: Email to be searched
          example: damian.obrien@gmail.com
    PageEmailResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/EmailResponse'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
...